backend/requirements.txt
flask
flask-cors
flask-mysqldb
flask-jwt-extended
python-dotenv
flask
flask-cors
flask-mysqldb
flask-jwt-extended
python-dotenv
pytest
pytest-flask
